removed the warning by adding implicit conversion -- EDColor/UIColor+Hex.m:67:25: Implicit conversion loses integer precision: 'unsigned long long' to 'int'

+++ EDColor/UIColor+Hex.m
@@ -64,7 +64,7 @@
     switch (strlen(s)) {
         case 2:
             // xx
-            r = g = b = value;
+            r = g = b = (int)value;
             a = 255;
             break;
         case 3:
